Allow returning a checked pointer from a function with an interop type. (#368)
Add support for returning checked pointers from functions with bounds-safe interface for their return values.   This addresses issue #366. The problem was that information about return bounds was not being propagated.  Modify SemaInit to propagate the return bounds information for the function to the EK_Result entities, which are used to typecheck return statements.   The existing machinery in SemaInit for handling C assignments takes care of the rest of the checking.

Testing:
- Added tests to the Checked C repo for returning checked values in  bodies of functions with bounds-safe interfaces for their return.  Modified  an existing test to allow this behavior.
- Passes existing Checked C repo and clang Checked C tests.
- Passes  automated testing for debug Windows x64 and Linux x64.