Fix incorrect logic for inserting bounds-safe interface cast. (#376)
This addresses Github issue #369.   We were incorrectly comparing values of different enum types when deciding whether to insert implicit casts at bounds-safe interfaces.  Change the comparison to be correct.
Rework the logic to follow the earlier code in the function for inserting casts. The earlier code has a logic twisted up like a pretzel, so we follow it exactly to avoid subtle problems.

If we are able to type check using the LHSInteropType, we insert a cast to the LHSInteropType, if necessary (CheckAssignmentConstraints takes the RHS by reference and may already have altered it).  Then we always insert a  bitcast from the LHSInteropType to the LHSType.

Testing:
- Passed local testing of Checked C repo and clang Checked C tests on Windows on x64.
- Passed automated testing for X64 Linux for clang tests and LNT.