Check that call arguments have bounds when expected. (#377)
 If a function parameter has a bounds declaration, check that the corresponding argument at a call has known bounds.   This addresses issue #373.

Also plumb in support for checking declared bounds at function calls.  We take the declared parameter bounds for a parameter and substitute the actual arguments to create an expected bounds for the argument.   For arguments for parameters used in bounds expressions, we can't handle modifying (side-effecting) expression. We will issue an error if we see this.   We then try to check that the inferred bounds for each argument expression imply the actual bounds.

Testing:
 - Add a new test file bounds_decl_checking.c to the Checked C repo  tests\static_checking directory.  This checks that expressions  have bounds when expected.
- Add tests to the clang\tests\CheckedC\bounds-decl-checking.c file (for -Wcheck-bounds-decl) that check that warnings are not issued for simple cases such as passing null or an argument whose expected bounds exactly match the bounds for the argument. 
- Update one of the convert LNT benchmarks to add a missing bounds declaration found by this checking.
- Pass automated testing for x64 Linux for the clang regression suite and LNT.