Use the declared bounds for an array if they are present.
If a variable with an array type or a member with an array type also has
declared bounds, use the declared bounds.  We have been using bounds just
based on the array type.

The logic for variables and member accesses is similar, except that the
member bounds need to beconcretized.

This allows provide bounds checking for variable-length arrays at the end of
structs:
struct S {
  int len;
  int arr checked[] : count(len);
}

Testing:
- A new file array-bounds-decls.c is being to the Checked C repo for testing
bounds checking for arrays with declare bounds.