Avoid duplicate error messages when T * is cast to ptr<T>.
This addresses issue #242.  There were duplicate messages when
a value of T * is cast to a ptr variable.  One error message
was generated for the cast; anoother one was generated for the
consumer of the cast.

The fix is to base the bounds of the cast expression on the target
type of the cast when the target type is a _Ptr.  This is the
correct fix because we should be narrowing bounds anyway.