Avoid duplicate error messages when T * is cast to ptr<T>.  (#385)
This addresses issue #242.  There were duplicate messages when
a value of T * is cast to a ptr variable.  One error message
was generated for the cast.  Another one was generated for the
consumer of the cast.

The fix is to determine the bounds of the cast expression by using
bounds inferred from the target type of the cast, when the target
type is a _Ptr.     This is the correct fix because we should be 
narrowing bounds anyway.

Testing:
- Added regression test case that checks that only one error message
  is issued.