Fix duplicate error message when initializer has an error. (#388)
When an initializer has a syntax or typechecking error, it is not added to the
AST.  This causes the compiler to complain during static checking of bounds
declarations about the lack of an initializer.   The fix is to issue the error message
earlier.  This addresses issue #386.

Also fix some line ending issues for the ptr_cast_bug_242.c test case.

Testing:
- Added a regression test case that checks that only one diagnostic message
  about the error in the initializer is emitted