Checked pointer types display incorrectly in aka clauses in diagnostics.
Checked pointer types were being displayed as unchecked pointer types
in "aka" clauses for types named by typedefs that are printed in
diagnostics.  The "aka" clause displays the underlying type that was
typedef'ed.

The fix is a one-line change in to pass the pointer type kind through
in the Desugar routine in ASTDiagnostics.  This change addresses issue #254.

Testing:
- Add regression test  that checks that checked types display
  properly in "aka" clauses.