Modify bounds checking for null-terminated pointers. (#399)
Checked C allows reads at exactly the upper bound for null-terminated pointers
so that the potential terminating value can be examined.  This modifies runtime
array bounds checks to allow that.

The IR is extended with a bounds check kind, so that we know when generating
a check whether the check should allow the upper bound location to be read.
Bounds checks are done as part of lvalue expression evaluation, so we need
to know whether the lvalue will be read or written.  We can't just look at the
type of the pointer to determine the kind of bounds check.  The bounds check
kind is added to the dereference and array subscript expression classes.
It is stored in the statement bitfields data structure, so no extra memory
is used.

For code generation of bounds checks for a null-terminated pointer read,
generate a 'less than or equal' comparison instead of a 'less than'
comparison.

Testing:
- Added new dynamic tests for the Checked C repo in nullterm_pointer.c that
  check that the memory location at the upper bound is accessible
  for null-terminated array pointers, and not accessible for
  non-null-terminated array pointers.
- Passing existing automated testing for Linux.