Track compiler-inferred bounds and fix bugs.
Change the IR so that we can track default bounds that are compiler
inserted.  Diagnostics involvings bounds mismatches are more complicated
to generate when some of the bounds are inferred because the bounds don't
appeaer int he program text.  In addition, when printing out types,
we want to suppress printing bounds that weren't declared in the original
source code.

Change GetCheckedCInteropType to return the same type when it sees
bounds(unknown).  We are now seeing bounds(unknown) in more places than
before.

Compute the default bounds properly for a parameter with check array
type.

Only check the bounds for an initializing expression when the variable
being initialized is a scalar type.  We don't need to check bounds when
the initializing expression is a non-scalar type (such as an array).
The correct amount of space for the type will always be allocated.

Still to do:
- Delete redundant code that is inserting bounds for parameters.
- Delete redundant code in bounds inference that sets default bounds
  for variables.
- Delete redundant code in function type creation that sets default bounds for
  parameters.
- Improve error messages involving default bounds.