Warn in checked scopes about unprovable bounds declarations.
The Checked C compiler will now always warn in checked scopes about bounds
declarations that cannot be proved true or false.  For unchecked scopes,
the warning is off by default.

I replaced the RecursiveASTVisitor with a recursive function because
I needed to pass information about checked scopes downward from
compound statements to their children.

To avoid having to unnecessarily change tests in the Checked C repo,
I generalized the code for constant-sized ranges to detect
equivalence of base expressions beyond lexical equivalence.  It
now canonicalizes base expressions by stripping off outer operations
that do not change the runtime value of the expression.  This includes
parentheses, value-preserving casts, and certain operations involving
function/array types.  For example, the address-of operator applied to an
array does not actually change runtime values.  It gives you a pointer to the
same location with a different type.

Testing:
- Updated Checked C repo tests with new warnings about that occur
for bounds declarations in checked scopes.  There will be a separate
pull request for this.
- For the clang Checked C repo, added a new test of bounds declarations that
verifies that warnings are always produced for checked scopes.  It includes
the existing test of bounds declarations, but uses a pragma to put it in a
checked scope.
- Update tests with new warning messages.