Allow writing null-terminators at nt_array_ptr upper bounds.
When converting existing code, a common pattern is to write a null-terminator
after copying data to an nt_array_ptr.  This can lead to attempting to write a
null-terminator exactly at the upper bound of the null-terminated array,
when the entire is filled with deta. This was not allowed under Checked C
rules.  It's legal to allow a write of a null-terminator at the upper
bound of a null-terminated array: the invariant for a null-terminated array
is still met.

This change adds support for writing a null-terminator at exactly the
upper bound of an null-termianted array.  We only allow this for simple
assignments.  It is not allowed for compound assigments.  This was a
little awkward to implement because all bounds checks are currently done
as part of lvalue-expression evaluation.   In this case, we need to know
the value that is being written.   I didn't want to move the
bounds check insertion to all places that use lvalues.  Instead, we
skip inserting the check when constructing an lvalue for this specific
case.   We have a special case check in assignment.  We might to revisit
this in the future and move the checks.

Testing:
- Added new tests to the nullterm_pointers.c file in the Checked C repo.
The tests check writing null values and non-null values exactly at the
upper bounds of nt_array_ptr and regular array_ptrs.