Correct nt_array_ptr write bounds check.
When we allow a write of 0 at the upper bound of a
pointer to a null-terminated array, we have to expand
the range of allowed memory to include the first element
of the null terminated sequence.  It is possible for
even this range to be empty.  If so, the write needs
to fail with a runtime fault.

This adjusts the bounds check for a write of 0 at the
upper bound to  also check the lower bound.  This prevents
a write to an empty expanded memory range.

Testing:
- Add tests to nullterm_pointers.c in the Checked C repo.
  The tests check writes of 0 at the upper bound where
  the expanded range has 1 element or no elements.