Delete redundant dyn_casts. (#429)
This is a follow up on PR #406.   @awruef noted that we were using dyn_cast when we should just use cast.   The issue is that while the dyn_casts may seem to make the code more reliable, they don't.   The dyn_cast implemention is using the same information that we just switched on.  This change removes the redundant dyn_casts.

Testing:
- Passed automated testing on Linux and Windows.