Allow writing null terminators at nt_array_ptr upper bounds. (#427)
In existing C code, a common pattern is to write a null terminator after copying data to a null-terminated array.  When converting the code to Checked C, this can lead to attempting to write a null-terminator exactly at the upper bound of an nt_array_ptr, when the entire array is filled with data. This was not allowed under Checked C bounds checking rules.   Even though it was not allowed, it is OK to allow it because the invariant for null-terminated arrays is preserved by the write. 

This change relaxes the bounds checking for nt_array_ptrs to allow writing null values exactly at upper bounds.  There is a matching Checked C repo change that updates the specification and tests.

We only allow this for simple assignments.  It is not allowed for compound assignments.  This was a little awkward to implement because all bounds checks are currently done as part of lvalue-expression evaluation.   In this case, we need to know the value that is being written.   I didn't want to move the bounds check insertion to all places that use lvalues.  Instead, we skip inserting the check when constructing an lvalue for this specific case.   We have a special case check in assignment.  We might to revisit this in the future and move the checks.

The generated LLVM IR for the check for the upper bound case is larger than necessary.   Given bounds (lb, ub), a destination pointer p,, and a value v, we check three conditions: p == ub &&  lb <= p and v == 0.     The lower bound check is needed to handle empty ranges properly.  We could avoid generating code for it by changing the regular bounds check to use branches instead of AND'ing conditions together.    That's a  significant change that should be evaluated separately because it could have performance effects.

Testing:
- Added new tests to the nullterm_pointers.c file in the Checked C repo.  The tests check writing null values and non-null values exactly at the upper bounds of nt_array_ptr and regular array_ptrs.
- Passes automated testing.