Automation for building install packages. (#434)
Extend the Windows automation scripts for VSTS to support building an install package.   This adds a new environment variable BUILD_PACKAGE that can be set to Yes or No.  It defaults to No.   When is true, an installation package is built and placed in the subdirectory package of the object directory.

Testing:
- Tested automation using the build-installer branch.  This built a Windows X64 installation package.