Enable Checked C extension by defaults.
Enable the Checked C extension by default when compiling C files.  It can
be turned off using the -fno-checkedc-extension flag.   The
-fcheckedc-extension flags remains too, though it doesn't do anything
for C files.

Fix bugs found by building the clang test suite with the Checked C extension
enabled.
- C11's generic construct and Checked C's return bounds expression lead to
 syntactic ambiguity during parsing, if you are only looking at the
 next token.  C11's generic construct allows cases of the from
 "type name : expr".  Checked C allowed return bounds expression
 on function types.  Type name could have the form "return-type
 function-name(... parameter list) : bounds expression.  The fix
 is to look at the token after the ":" to see if it starts a bounds
 expression.
- Call targets in call expressions can have block pointer types, in
  addition to function pointer types.
- We did not allow clang's "block" extension when processing return
  bounds for functions.  There is no syntax for writing a return bounds
  on a block, so it is fine to allow them.
- The code for skipping an invalid bounds expression did not handle
end-of-file tokens properly, cause a crash if a bounds expression was missing
at the end of a pre-processed file.
- The expected source location for a missing bounds expresion was confusing.
The code printed the starting location of the next token, which could be on
another line and might have nothing to do with bounds expressions.
A better approach that is used for missing expected tokens elsewhere in
clang to use the end of the current token.  The end is not inclusive,
so it is 1 character past the token.
- For a redeclaration of function where the declaration using old-style
K&R parameter syntax, if there was a type mismatch, the location of the
prior declaration was printed twice, including one before the actual
error message.   The problem was that we have a method that analyzes a
type mismatch to see if it was caused by bounds information.  It
was always printing the location of the prior declaration, even if the
type mismatch was not caused by mismatched bounds information.

Also fix a bug in the Checked C rewriter: the code wanted to use interface
type information.  The code was always fetching the type of the bounds
expressions, assuming that given the checked type for used in an interface.
Bounds expressions have no type.  InteropTypeBoundsAnnotations, which
are stored in the same location as bounds expressions, carry the
interface type as their type.