Fix compiler assertion at bounds-safe interface assignment. (#438)
Clang's code generation phase assumes for lvalue casts that (1) the cast must be explicit and (2) the explicit cast type has the type "reference to T".   It assumes that lvalue casts arise from reinterpret casts in C++ of the form `reinterpret_cast<T&>`.

This updates the code generation phase to handle the implicit lvalue cast for bounds-safe interfaces.  One subtle aspect is that in this case, the cast just has a target type of "checked pointer to S", not "reference to checked pointer to S".  When building the LLVM type of the address of the lvalue, we have to start with a "pointer to a checked pointer to S". 

It's possible that we should change the type of the cast.   If I try to change the type of the cast for all uses of lvalues with a bounds-safe interface,  this triggers an assert when the lvalue is used within value-producing expressions.  It's a bit odd because we should proceed to insert an lvalue-to-rvalue cast in that situation, so it shouldn't be a problem.  I've opened issue #438 to track this.

Testing:
- Add new to the Checked C repo that tests the failure case.  That will be in a separate PR.
- Passed existing automated testing on Linux.
- Passed a local clang testing on x64 Windows.