Fix bug with declaration context lookup. (#450)
When I added logic to change variable lookup to be sensitive to whether or not you wanted the parameter or the global declaration, I didn't thread through this logic into the recursive calls of `getVariable`. So sometimes you ask for the parameter but instead you get the global. This fixes that by threading through the boolean flag indicating whether or not you want the parameter or the global.