Fix subtle issues related to substitution of arguments for parameters (#453)
This addresses issue #451.  During checking of bounds declarations, we need to check that the actual arguments meet the bounds constraints of the parameters.  The way we do this is by taking the bounds declaration for a parameter and substituting in the actual arguments to derive the desired bounds expression, in the context of the caller.  We then check that the actual argument implies the validity of this bounds expression.

It turns out that substituting an argument expression for a parameter variable is trickier than you might think in clang.  The issue is that the argument expression might be subject to implicit casts.   You can substitute the expression in for the parameter in your bounds expression, but if you call TreeTransform on the bounds expression, your implicit casts might get dropped.    The rules for implicit casts differ between call arguments/assignments and other operators.   This can lead to changes in the types of subexpressions of the bounds expressions, which could change their meaning or lead to illegally-typed expressions.  For example, an argument expression with void pointer type could be implicitly cast to some pointer to T.   The implicit cast could be lost, leading to a bounds expression with pointer arithmetic on void pointers, which is nonsensical.

For now, I'm taking the approach that we'll convert these implicit casts to explicit casts in the IR, so that these subtle problems don't happen.  Note that we can't keep all implicit casts during TreeTransform: it assumes that implicit casts are being resynthesized.

- This change updates argument substitution to convert implicit casts that aren't the usual unary conversions to explicit casts. 
- It also removes a version of GetTypeBasedBounds that can be dangerous to use.   The bounds inference code is generally expecting to work on expressions in an "expanded" form, where count/byte_count have been removed as sugar.   The removed version of GetTypeBasedBounds produces a count/byte_count, which could unexpectedly propagate if the caller isn't careful.
- It puts the target bounds expressions for a parameter in the expanded standard form as soon as possible, when checking call arguments.  

This change also fixes a buggy use of getCurScope() during semantic checking of C-style casts.   getCurScope was used to see if the operator is in a checked scope. However, the getCurScope method should only be used during arsing.  It should not used in code that may be invoked elsewhere in the compiler, such as TreeTransform.   Our use of TeeTransform for substitution sometimes failed because we had introduced explicit casts that would be disallowed at source-level in a checked scope (specifically for implicit casts for bounds-safe interfaces).

I modified semantic checking of CStyleCasts to only verify the Checked C requirements for C-style casts in checked scopes during parsing.  This is only a partial solution.  A better solution would be record whether a CStyleCast operation occurs in a checked scope when it is created, so that template substitution works properly.   That change would be mostly unrelated to this change  so it should be done separately.

Testing:
- Passed local testing on Windows.
- Passed automated testing on Linux, including running the LNT tests and tests that have been converted to Checked C.
- Added a new test that highlights the subtle issue around a dropped conversion.