Fix compiler assert when using pointers with null-terminated interface types. (#456)
This change fixes issue #449.  You can use the `itype` construct to specify that a pointer has an interface type that is a null-terminated pointer type: `char *p : itype(nt_array_ptr<char>)'.   The null-terminated pointer will have bounds of count(0), which means it can be read and written (only null values can be written).

Reading or writing via the pointer triggered an internal compiler assert.  The problem was that we were creating an lvalue-to-rvalue cast where the source and destination types were not equal (they differed in checkedness).  The fix is to use the same type for the lvalue-to-rvalue cast, and cast the resulting value.

This fix produces some chains of casts in error messages involving bounds declaration.  These were full of redundancy and hard to understand.   When creating an explicit cast, elide intermediate casts that don't change the value of an expression.  Use the same method we use during semantic analysis of bounds declarations.   This makes the (bounds) ranges described in error messages simpler and easier to use.

Also make sure that Ptr values are cast to Array_ptr when putting bounds in standard form.  We were expanded bounds to ranges that were doing pointer arithmetic on Ptr values.   When TreeTransform is run, it re-runs semantic analysis, which errors on those ranges.

Testing:
- Passed local testing on Windows x64
- Passed automating testing on Linux x64, including LNT.
- Add new regression test to the Checked C repo that tests that dynamic behavior is what is expected.  This was covered by a  separate pull request  (https://github.com/Microsoft/checkedc/pull/259).