Fix compiler assert involving dynamic_bounds_cast from ptr<void>.
This fixes issue #458, a user-reported compiler crash.  Some simple code that
converted from a ptr<void> to ptr<char> caused an internal compiler crash.

We were generating malformed IR for this operation.  I found two problems:
1. We have code for deleting unnecessary C-style cast operations during the
inference of bounds information.  This codes clean up the bounds so that
compiler error-message are easier to read.  We were being too aggressive
in deleting casts, resulting in invalid casts at the LLVM (we deleted an
array-to-pointer decay cast).  The fix is to be less aggressive.  I created a
separate method for determining whether a cast should be removed, independent
of the current method that determines whether a cast does not change values
(and can be ignored when comparing expressions for equality).  This issue
is what likely caused the crash.
2. We were not properly constructing the target (desired) bounds for the
runtime checking.  Given dynamic_bounds_cast<ptr<T>>(e1), we were putting
target bounds for "e1 : count(1)" into a normalized form.   To ensure the
bounds are wide enough for a value of type T, the bounds to be at the
type ptr<T>, not whatever type e1 happens to be.  So we need to put
((ptr<T>) e1) : count(1) into a normalized form.  Also, if ptr<T> is ptr<void>,
we need to use byte_count(1).

This change includes several clean ups:
- In the representation of cast expressions, we had some methods of the form
{get,set}CastBounds, as well as an associated enum value for the position of
the operand.  This isn't very clear and is confusing to read.  This is a
normalized form of the bounds expression, which we create for the benefit
of the code generator.  Rename the methods and values.
- The code for ActOnBoundsCastExprSingle handles the case for
dynamic_bounds_cast<ptr<T>>. It was constructing a count(1) expression.
We now have count(1) and byte_count(1) prebuilt, so use that instead.  We also
had some code for attaching a relative alignment clause.  This is for obsolete
syntax; issue #448 covers removing  the syntax.  In fact, this code never did
anything.   The bounds expression is always a count bounds expression.
It never made sense to allow a relative alignment clause for this case and
there should have been an error message generated.  We're going to delete this
syntax anyway, so delete the useless code (continue throwing away the clause.)\